+++ CHANGELOG.md
@@ -10,7 +10,7 @@ This is a small maintenance release.
 
 ### Changelog
 
-- Feat: New custom Lua function: mympd.splitlines(), mympd.trim()
+- Feat: New custom Lua function: mympd.splitlines(), mympd.trim(), mympd.read_file()
 - Upd: Translations
 - Upd: Mongoose to current master